+++ sigle/src/config.ts
@@ -14,6 +14,7 @@ export const sigleConfig = {
   messengerUrl: 'https://m.me/sigleapp',
   documentationUrl: 'https://docs.sigle.io',
   feedbackUrl: 'https://sigle.canny.io/feature-requests',
+  blogUrl: 'https://app.sigle.io/sigleapp.id.blockstack',
   email: 'sigle@protonmail.com',
   // This reflect the tailwindcss config
   breakpoints: {