+++ sigle/src/pages/index.tsx
@@ -3,19 +3,27 @@ import { useRouter } from 'next/router';
 import { Protected } from '../modules/auth/Protected';
 import { Home } from '../modules/home';
 import {
-  enableExperimentalEditor,
-  isExperimentalEditorEnabled,
-} from '../modules/editor/utils/experimentalEditor';
+  enableExperimentalHiroWallet,
+  isExperimentalHiroWalletEnabled,
+  isExperimentalThemeToggleEnabled,
+  enableExperimentalThemeToggle,
+} from '../utils/featureFlags';
 
 const HomePage = () => {
   const router = useRouter();
-  const isExperimentalEditor = router.query.experimentalEditor === 'true';
+  const isExperimentalHiroWallet =
+    router.query.experimentalHiroWallet === 'true';
+  const isExperimentalThemeToggle =
+    router.query.experimentalThemeToggle === 'true';
 
   useEffect(() => {
-    if (isExperimentalEditor && !isExperimentalEditorEnabled) {
-      enableExperimentalEditor();
+    if (isExperimentalHiroWallet && !isExperimentalHiroWalletEnabled) {
+      enableExperimentalHiroWallet();
     }
-  }, [isExperimentalEditor]);
+    if (isExperimentalThemeToggle && !isExperimentalThemeToggleEnabled) {
+      enableExperimentalThemeToggle();
+    }
+  }, [isExperimentalHiroWallet]);
 
   return (
     <Protected>