+++ sigle/src/ui/HoverCard.ts
@@ -1,0 +1,57 @@
+import * as HoverCardPrimitive from '@radix-ui/react-hover-card';
+import { keyframes, styled } from '../stitches.config';
+
+const slideUpAndFade = keyframes({
+  '0%': { opacity: 0, transform: 'translateY(2px)' },
+  '100%': { opacity: 1, transform: 'translateY(0)' },
+});
+
+const slideRightAndFade = keyframes({
+  '0%': { opacity: 0, transform: 'translateX(-2px)' },
+  '100%': { opacity: 1, transform: 'translateX(0)' },
+});
+
+const slideDownAndFade = keyframes({
+  '0%': { opacity: 0, transform: 'translateY(-2px)' },
+  '100%': { opacity: 1, transform: 'translateY(0)' },
+});
+
+const slideLeftAndFade = keyframes({
+  '0%': { opacity: 0, transform: 'translateX(2px)' },
+  '100%': { opacity: 1, transform: 'translateX(0)' },
+});
+
+const StyledContent = styled(HoverCardPrimitive.Content, {
+  cursor: 'pointer',
+  borderRadius: '$1',
+  py: '$2',
+  px: '$5',
+  backgroundColor: '$gray3',
+  color: '$gray11',
+
+  '&:hover': {
+    color: '$orange11',
+  },
+
+  '@media (prefers-reduced-motion: no-preference)': {
+    animationDuration: '400ms',
+    animationTimingFunction: 'cubic-bezier(0.16, 1, 0.3, 1)',
+    animationFillMode: 'forwards',
+    willChange: 'transform, opacity',
+    '&[data-state="open"]': {
+      '&[data-side="top"]': { animationName: slideDownAndFade },
+      '&[data-side="right"]': { animationName: slideLeftAndFade },
+      '&[data-side="bottom"]': { animationName: slideUpAndFade },
+      '&[data-side="left"]': { animationName: slideRightAndFade },
+    },
+  },
+});
+
+const StyledArrow = styled(HoverCardPrimitive.Arrow, {
+  fill: '$gray3',
+});
+
+export const HoverCard = HoverCardPrimitive.Root;
+export const HoverCardTrigger = HoverCardPrimitive.Trigger;
+export const HoverCardContent = StyledContent;
+export const HoverCardArrow = StyledArrow;